+++ saleor/checkout/models.py
@@ -9,6 +9,7 @@ from django.core.validators import MinValueValidator
 from django.db import models
 from django.utils.encoding import smart_str
 from django.utils.translation import pgettext_lazy
+from django_countries.fields import Country, CountryField
 from django_prices.models import MoneyField
 from prices import Money
 
@@ -42,6 +43,10 @@ class CheckoutQueryset(models.QuerySet):
             "lines__variant__product__images",
             "lines__variant__product__product_type__product_attributes__values",
         )  # noqa
+
+
+def get_default_country():
+    return settings.DEFAULT_COUNTRY
 
 
 class Checkout(ModelWithMetadata):
@@ -78,6 +83,7 @@ class Checkout(ModelWithMetadata):
         max_length=settings.DEFAULT_CURRENCY_CODE_LENGTH,
         default=settings.DEFAULT_CURRENCY,
     )
+    country = CountryField(default=get_default_country)
 
     discount_amount = models.DecimalField(
         max_digits=settings.DEFAULT_MAX_DIGITS,
@@ -142,6 +148,27 @@ class Checkout(ModelWithMetadata):
     def get_last_active_payment(self) -> Optional["Payment"]:
         payments = [payment for payment in self.payments.all() if payment.is_active]
         return max(payments, default=None, key=attrgetter("pk"))
+
+    def set_country(
+        self, country_code: str, commit: bool = False, replace: bool = True
+    ):
+        """Set country for checkout."""
+        if not replace and self.country is not None:
+            return
+        self.country = Country(country_code)
+        if commit:
+            self.save(update_fields=["country"])
+
+    def get_country(self):
+        address = self.shipping_address or self.billing_address
+        saved_country = self.country
+        if address is None or not address.country:
+            return saved_country.code
+
+        country_code = address.country.code
+        if not country_code == saved_country.code:
+            self.set_country(country_code, commit=True)
+        return country_code
 
 
 class CheckoutLine(models.Model):