+++ saleor/core/management/commands/populatedb.py
@@ -18,6 +18,7 @@ from ...utils.random_data import (
     create_shipping_zones,
     create_users,
     create_vouchers,
+    create_warehouses,
     set_homepage_collection,
 )
 
@@ -88,6 +89,8 @@ class Command(BaseCommand):
         create_images = not options["withoutimages"]
         for msg in create_shipping_zones():
             self.stdout.write(msg)
+        create_warehouses()
+        self.stdout.write("Created warehouses")
         create_products_by_schema(self.placeholders_dir, create_images)
         self.stdout.write("Created products")
         for msg in create_product_sales(5):