+++ saleor/graphql/account/schema.py
@@ -27,6 +27,7 @@ from .mutations.account import (
     RequestEmailChange,
 )
 from .mutations.base import (
+    ConfirmAccount,
     PasswordChange,
     RequestPasswordReset,
     SetPassword,
@@ -228,6 +229,7 @@ class AccountQueries(graphene.ObjectType):
 class AccountMutations(graphene.ObjectType):
     # Base mutations
     request_password_reset = RequestPasswordReset.Field()
+    confirm_account = ConfirmAccount.Field()
     set_password = SetPassword.Field()
     password_change = PasswordChange.Field()
     request_email_change = RequestEmailChange.Field()