+++ saleor/graphql/core/types/common.py
@@ -14,6 +14,8 @@ from ..enums import (
     ProductErrorCode,
     ShippingErrorCode,
     ShopErrorCode,
+    StockErrorCode,
+    WarehouseErrorCode,
     WebhookErrorCode,
     WishlistErrorCode,
 )
@@ -84,6 +86,14 @@ class GiftCardError(Error):
 
 class ExtensionsError(Error):
     code = ExtensionsErrorCode(description="The error code.")
+
+
+class StockError(Error):
+    code = StockErrorCode(description="The error code.")
+
+
+class WarehouseError(Error):
+    code = WarehouseErrorCode(description="The error code.")
 
 
 class WebhookError(Error):