+++ saleor/graphql/product/mutations/products.py
@@ -887,13 +887,9 @@ class ProductCreate(ModelMutation):
             track_inventory = cleaned_input.get(
                 "track_inventory", site_settings.track_inventory_by_default
             )
-            quantity = cleaned_input.get("quantity", 0)
             sku = cleaned_input.get("sku")
             models.ProductVariant.objects.create(
-                product=instance,
-                track_inventory=track_inventory,
-                sku=sku,
-                quantity=quantity,
+                product=instance, track_inventory=track_inventory, sku=sku
             )
 
         attributes = cleaned_input.get("attributes")