+++ tests/test_core.py
@@ -125,12 +125,12 @@ def test_create_fake_users(db):
 
 
 def test_create_address(db):
-    assert Address.objects.all().count() == 0
+    assert not Address.objects.exists()
     random_data.create_address()
     assert Address.objects.all().count() == 1
 
 
-def test_create_fake_order(db, monkeypatch, image, media_root):
+def test_create_fake_order(db, monkeypatch, image, media_root, warehouse):
     # Tests shouldn't depend on images present in placeholder folder
     monkeypatch.setattr(
         "saleor.core.utils.random_data.get_image", Mock(return_value=image)